% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockup.R
\name{mocktwit}
\alias{mocktwit}
\title{Create a mockup of a Twitter status}
\usage{
mocktwit(x, file = NULL)
}
\arguments{
\item{x}{Twitter status data frame (as returned by rtweet).}

\item{file}{File name to save as. Defaults to temporary file.}
}
\value{
Saves an html file
}
\description{
Generates HTML twitter status page.
}
