% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockup-status.R
\name{mocktwitter_status}
\alias{mocktwitter_status}
\title{Create a mockup of a Twitter status}
\usage{
mocktwitter_status(x, file = NULL)
}
\arguments{
\item{x}{Twitter status data frame (as returned by rtweet).}

\item{file}{File name to save as. Defaults to temporary file.}
}
\value{
Saves an html file
}
\description{
Generates HTML twitter status page.
}
\examples{
\dontrun{
## create mock-up of Twitter status about rtweet release using

## the status URL
mocktwitter_status("https://twitter.com/kearneymw/status/1009431823791902720")

## the status_id
mocktwitter_status("1009431823791902720")

## or pick a tweet from my timeline to display
kmw <- rtweet::get_timeline("kearneymw")
mocktwitter_status(kmw[1, ])

}
}
