% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockup-timeline.R
\name{mocktwitter_timeline}
\alias{mocktwitter_timeline}
\title{Create a mockup of a Twitter timeline}
\usage{
mocktwitter_timeline(x, file = NULL)
}
\arguments{
\item{x}{Twitter timeline data frame (as returned by rtweet).}

\item{file}{File name to save as. Defaults to temporary file.}
}
\value{
Saves an html file
}
\description{
Generates HTML twitter timeline page.
}
\examples{
\dontrun{
## create mock-up of Twitter timeline

## the timeline URL
mocktwitter_timeline("https://twitter.com/kearneymw")

## the screen name or user ID
mocktwitter_timeline("kearneymw")

## or get timeline data from rtweet
kmw <- rtweet::get_timeline("kearneymw")
mocktwitter_timeline(kmw)

}
}
